requirements-dev.txt
black
ruff
pytest
fastapi
uvicorn[standard]
pydantic
pytest-jsonschema
jsonschema
PyYAML
